refactor(profile): rename ActivityDrawer to ActivityDrawerHost
Match the host-pattern convention introduced with MatchDrawerHost: the
host owns data fetching and renders the actual <Drawer>, while *Drawer
names are reserved for pure-content drawers.
